separates context migrate from agent file for readability
